+++ example/lib/main.dart
@@ -35,6 +35,7 @@ class HomePage extends StatelessWidget {
               onSelect: (Currency currency) {
                 print('Select currency: ${currency.name}');
               },
+              favorite: ['SEK'],
             );
           },
           child: const Text('Show currency picker'),